Remove STATIC_URL template variable
STATIC_URL has always simply been `/static/`. We've not so far had
cause to change it, and yet it is an extra level of complexity for
developers to understand and for the application to implement.

If `/static/` ever changes location, doing a global find/replace
isn't that hard. So I don't think it's needed. Remove it.